cloudwatch logging with the `metric_logger` conventions
